Refactor sugerencias toast
